Made hf-tokenizers optional, added doc for HFTokenizer
